4 Задание
